Create jupyterdiffusion.ipynb
This is an adaptation of the **RFdiffusion** Colab Implementation to run locally in Jupyter Notebook, preferably on the LRZ AI Cluster. It does not require any prior installations, you just have to keep in mind that all data is going to be saved in your home directory in a new directory called \"RFdiffusionJupyter\". So if you have RFdiffusion installed already you could for example save a little disk space by combining this directory with the original RFdiffusion directory.